Make main.py into __main__.py
